requirements.txt
streamlit
streamlit-extras
python-dotenv
pypdf2
langchain
sentence-transformers
faiss-cpu